pyproject.toml
[project]
name = "my-mtproto-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "celery>=5.6.2",
    "django>=5.2.11",
    "djangorestframework>=3.16.1",
    "factory-boy>=3.3.3",
    "gunicorn>=25.1.0",
    "pytelegrambotapi>=4.31.0",
    "redis>=7.2.1",
    "requests>=2.32.5",
]
